Fix incompatibility with tf.data.Dataset
When you try to use this callback with a dataset, it errors out later because `samples` has a value of `None`.  However, there is a value for `steps`, so the else does take care of the dataset case. This change assures that the dataset case is directed to the else.